requirements.txt
tqdm
unidecode
pillow
